Update jwt security process

+++ src/main/java/org/seoultech/fitnesspotential/domain/community/controller/CommunityCommentController.java
@@ -3,10 +3,13 @@ package org.seoultech.fitnesspotential.domain.community.controller;
 import org.seoultech.fitnesspotential.domain.community.dto.CommunityCommentPostRequest;
 import org.seoultech.fitnesspotential.domain.community.dto.CommunityCommentPutRequest;
 import org.seoultech.fitnesspotential.domain.community.entity.Community;
+import org.seoultech.fitnesspotential.domain.community.entity.CommunityComment;
 import org.seoultech.fitnesspotential.domain.community.service.CommunityCommentService;
+import org.seoultech.fitnesspotential.domain.user.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,23 +25,28 @@ public class CommunityCommentController {
     }
 
     @PostMapping
-    public ResponseEntity<Community> postCommunityComment(@RequestBody @Validated CommunityCommentPostRequest communityCommentPostRequest) {
-        Community community = communityCommentService.postCommunityComment(communityCommentPostRequest, 0L); // When social login issue done, replace creatorId
-
+    public ResponseEntity<Community> postCommunityComment(@RequestBody @Validated CommunityCommentPostRequest communityCommentPostRequest, @RequestAttribute(name = "user") User user) {
+        Community community = communityCommentService.postCommunityComment(communityCommentPostRequest, user.getId()); // When social login issue done, replace creatorId
         return new ResponseEntity<>(community, HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Community> putCommunityComment(@RequestBody @Validated CommunityCommentPutRequest communityCommentPutRequest, @PathVariable Long id) {
+    public ResponseEntity<Community> putCommunityComment(@RequestBody @Validated CommunityCommentPutRequest communityCommentPutRequest, @PathVariable Long id, @RequestAttribute(name = "user") User user) {
+        CommunityComment communityComment = communityCommentService.getCommunityComment(id);
+        if (communityComment.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
         Community community = communityCommentService.putCommunityComment(communityCommentPutRequest, id);
-
         return new ResponseEntity<>(community, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCommunityComment(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCommunityComment(@PathVariable Long id, @RequestAttribute(name = "user") User user) {
+        CommunityComment communityComment = communityCommentService.getCommunityComment(id);
+        if (communityComment.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
         communityCommentService.deleteCommunityComment(id);
-
         return new ResponseEntity<>(null, HttpStatus.ACCEPTED);
     }
 }
