+++ src/main/java/org/seoultech/fitnesspotential/domain/community/controller/CommunityController.java
@@ -5,15 +5,19 @@ import org.seoultech.fitnesspotential.domain.community.dto.CommunityPutRequest;
 import org.seoultech.fitnesspotential.domain.community.entity.Community;
 import org.seoultech.fitnesspotential.domain.community.entity.CommunitySummary;
 import org.seoultech.fitnesspotential.domain.community.service.CommunityService;
+import org.seoultech.fitnesspotential.domain.user.entity.User;
+import org.seoultech.fitnesspotential.domain.user.exception.UserErrorMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 @RestController
@@ -42,19 +46,27 @@ public class CommunityController {
     }
 
     @PostMapping
-    public ResponseEntity<Community> postCommunity(@RequestBody @Validated CommunityPostRequest communityPostRequest) {
-        return new ResponseEntity<>(communityService.postCommunity(communityPostRequest, 0L), HttpStatus.OK);
+    public ResponseEntity<Community> postCommunity(@RequestBody @Validated CommunityPostRequest communityPostRequest, @RequestAttribute(name = "user") User user) {
+        return new ResponseEntity<>(communityService.postCommunity(communityPostRequest, user.getId()), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Community> putCommunity(@RequestBody @Valid CommunityPutRequest communityPutRequest, @PathVariable Long id) {
+    public ResponseEntity<Community> putCommunity(@RequestBody @Valid CommunityPutRequest communityPutRequest, @PathVariable Long id, @RequestAttribute(name = "user") User user) {
+        Community community = communityService.getCommunity(id);
+        if (community.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
         return new ResponseEntity<>(communityService.putCommunity(communityPutRequest, id), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCommunity(@PathVariable Long id) {
-        communityService.deleteCommunity(id);
+    public ResponseEntity<Void> deleteCommunity(@PathVariable Long id, @RequestAttribute(name = "user") User user) {
+        Community community = communityService.getCommunity(id);
+        if (community.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
 
+        communityService.deleteCommunity(id);
         return new ResponseEntity<>(null, HttpStatus.ACCEPTED);
     }
 }
