+++ src/main/java/org/seoultech/fitnesspotential/domain/fitness/controller/FitnessDiaryController.java
@@ -5,9 +5,11 @@ import org.seoultech.fitnesspotential.domain.fitness.dto.diary.FitnessDiaryPostR
 import org.seoultech.fitnesspotential.domain.fitness.dto.diary.FitnessDiaryPutRequest;
 import org.seoultech.fitnesspotential.domain.fitness.entity.FitnessDiary;
 import org.seoultech.fitnesspotential.domain.fitness.service.FitnessDiaryService;
+import org.seoultech.fitnesspotential.domain.user.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -30,32 +32,39 @@ public class FitnessDiaryController {
     }
     
     @GetMapping("/info")
-    public ResponseEntity<Iterable<FitnessDiaryInfo>> getFitnessDiaryInfos() {
-        return new ResponseEntity<>(fitnessDiaryService.getFitnessDiaryInfos(0L), HttpStatus.OK);
+    public ResponseEntity<Iterable<FitnessDiaryInfo>> getFitnessDiaryInfos(@RequestAttribute(name = "user") User user) {
+        return new ResponseEntity<>(fitnessDiaryService.getFitnessDiaryInfos(user.getId()), HttpStatus.OK);
     }
 
     @GetMapping("/search")
-    public ResponseEntity<Iterable<FitnessDiary>> getFitnessDiaries(@RequestParam Integer year, @RequestParam Integer month, @RequestParam Integer day) {
-        Iterable<FitnessDiary> fitnessDiaries = fitnessDiaryService.getFitnessDiaries(year, month, day, 0L);
+    public ResponseEntity<Iterable<FitnessDiary>> getFitnessDiaries(@RequestParam Integer year, @RequestParam Integer month, @RequestParam Integer day, @RequestAttribute(name = "user") User user) {
+        Iterable<FitnessDiary> fitnessDiaries = fitnessDiaryService.getFitnessDiaries(year, month, day, user.getId());
         return !fitnessDiaries.iterator().hasNext() ?
                 new ResponseEntity<>(fitnessDiaries, HttpStatus.NO_CONTENT) :
                 new ResponseEntity<>(fitnessDiaries, HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<FitnessDiary> postFitnessDiary(@RequestBody @Valid FitnessDiaryPostRequest fitnessDiaryPostRequest) {
+    public ResponseEntity<FitnessDiary> postFitnessDiary(@RequestBody @Valid FitnessDiaryPostRequest fitnessDiaryPostRequest, @RequestAttribute(name = "user") User user) {
 
-        return new ResponseEntity<>(fitnessDiaryService.postFitnessDiary(fitnessDiaryPostRequest, 0L), HttpStatus.OK);
+        return new ResponseEntity<>(fitnessDiaryService.postFitnessDiary(fitnessDiaryPostRequest, user.getId()), HttpStatus.OK);
     }
 
     @PutMapping
-    public ResponseEntity<FitnessDiary> putFitnessDiary(@RequestBody @Valid FitnessDiaryPutRequest fitnessDiaryPutRequest, Long id) {
-
+    public ResponseEntity<FitnessDiary> putFitnessDiary(@RequestBody @Valid FitnessDiaryPutRequest fitnessDiaryPutRequest, Long id, @RequestAttribute(name = "user") User user) {
+        FitnessDiary fitnessDiary = fitnessDiaryService.getFitnessDiary(id);
+        if (fitnessDiary.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
         return new ResponseEntity<>(fitnessDiaryService.putFitnessDiary(fitnessDiaryPutRequest, id), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteFitnessDiary(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteFitnessDiary(@PathVariable Long id, @RequestAttribute(name = "user") User user) {
+        FitnessDiary fitnessDiary = fitnessDiaryService.getFitnessDiary(id);
+        if (fitnessDiary.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
         fitnessDiaryService.deleteFitnessDiary(id);
         return new ResponseEntity<>(null, HttpStatus.ACCEPTED);
     }
