+++ src/main/java/org/seoultech/fitnesspotential/domain/fitness/controller/FitnessInfoController.java
@@ -6,12 +6,14 @@ import org.seoultech.fitnesspotential.domain.fitness.entity.FitnessInfo;
 import org.seoultech.fitnesspotential.domain.fitness.entity.FitnessInfoSummary;
 import org.seoultech.fitnesspotential.domain.fitness.service.FitnessInfoService;
 import org.seoultech.fitnesspotential.domain.fitness.service.FitnessInfoSummaryService;
+import org.seoultech.fitnesspotential.domain.user.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -45,19 +47,25 @@ public class FitnessInfoController {
     }
 
     @PostMapping
-    public ResponseEntity<FitnessInfo> postFitnessInfo(@RequestBody @Valid FitnessInfoPostRequest fitnessInfoPostRequest) {
-
-        return new ResponseEntity<>(fitnessInfoService.postFitnessInfo(fitnessInfoPostRequest, 0L), HttpStatus.OK);
+    public ResponseEntity<FitnessInfo> postFitnessInfo(@RequestBody @Valid FitnessInfoPostRequest fitnessInfoPostRequest, @RequestAttribute(name = "user") User user) {
+        return new ResponseEntity<>(fitnessInfoService.postFitnessInfo(fitnessInfoPostRequest, user.getId()), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<FitnessInfo> putFitnessInfo(@RequestBody @Valid FitnessInfoPutRequest fitnessInfoPutRequest, @PathVariable Long id) {
-
+    public ResponseEntity<FitnessInfo> putFitnessInfo(@RequestBody @Valid FitnessInfoPutRequest fitnessInfoPutRequest, @PathVariable Long id, @RequestAttribute(name = "user") User user) {
+        FitnessInfo fitnessInfo = fitnessInfoService.getFitnessInfo(id);
+        if (fitnessInfo.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
         return new ResponseEntity<>(fitnessInfoService.putFitnessInfo(fitnessInfoPutRequest, id), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteFitnessInfo(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteFitnessInfo(@PathVariable Long id, @RequestAttribute(name = "user") User user) {
+        FitnessInfo fitnessInfo = fitnessInfoService.getFitnessInfo(id);
+        if (fitnessInfo.getCreatorId() != user.getId()) {
+            throw new AccessDeniedException("Unauthorized");
+        }
         fitnessInfoService.deleteFitnessInfo(id);
         return new ResponseEntity<>(null, HttpStatus.ACCEPTED);
     }
